Change metric fetchers to fulfil an interface
`pkg/metric_endpoint` can act as a Gin request handler for a generic
metrics endpoint (any CF service.) It then calls a metric fetcher,
which is service-specific (e.g., for Redis.) This commit improves how
metric fetchers are structured. Instead of one function they are now
an interface. This gives more freedom and will be more fun to test.