Only fetch one metric datapoint
Before this commit I was experimenting with returning one metric
datapoint per minute and exposing all of them to Prometheus with
different timestamps. In practice getting Prometheus to not discard
the data was a pain.

This commit drops that idea. You'll get an average, minimum and
maximum of every metric across the last 5 minutes, and not the
individual minutely values. That's enough for now.

It's a shame that Prometheus won't arbitrarily import timestamped data
but that is a long-running controversy.