Add a (fake) integration test pipeline
This commit is to support using the build Concourse to add version
tags to this git repo. The "integration test" pipeline template
expects the file added by this commit to exist and be a runnable
Concourse task.

We don't have any integration tests for this repo, so the Concourse
task merely runs the build and unit tests. We've done similar things
in other repos before such as `paas-trusted-people`.